Fix #25 Fix field corners issues re-introduced by new bootstrap version Disable lastpass and dashlane where not needed
