+++ .gitconfig
@@ -1,0 +1,2 @@
+[core]
+	ignorecase = false
