+++ global-scoreboard/package.json
@@ -1,6 +1,5 @@
 {
   "name": "global-scoreboard",
-  "version": "0.1.0",
   "description": "An unofficial speedrun.com global scoreboard for friendly competitive speedrunning",
   "private": true,
   "scripts": {
@@ -11,14 +10,14 @@
     "lint": "eslint src/ --ext .ts,.tsx --ignore-path ../.gitignore",
     "lint:fix": "npm run lint -- --fix",
     "upgrade": "npx npm-check-updates -u & npm i",
-    "postinstall": "patch-package"
+    "postinstall": "patch-package & npm audit fix"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
-    "bootstrap": "^5.0.0",
+    "bootstrap": "^5.0.2",
     "http-status-codes": "^2.1.4",
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.0-alpha.2",
@@ -28,16 +27,17 @@
     "react-bootstrap-table2-toolkit": "^2.1.3",
     "react-dom": "^17.0.2",
     "react-picky": "^5.3.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.3"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.23",
-    "@types/node": "^15.12.2",
-    "@types/react": "^17.0.9",
-    "@types/react-dom": "^17.0.6",
-    "@typescript-eslint/eslint-plugin": "^4.26.1",
-    "@typescript-eslint/parser": "^4.26.1",
-    "eslint": "^7.28.0",
+    "@types/node": "^15.12.4",
+    "@types/react-router-dom": "^5.1.7",
+    "@types/react": "^17.0.11",
+    "@types/react-dom": "^17.0.8",
+    "@typescript-eslint/eslint-plugin": "^4.28.0",
+    "@typescript-eslint/parser": "^4.28.0",
+    "eslint": "^7.29.0",
     "eslint-plugin-extra-rules": "0.0.0-development",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prefer-arrow": "^1.2.3",
@@ -47,8 +47,9 @@
     "eslint-plugin-sonarjs": "^0.7.0",
     "eslint-plugin-switch-case": "^1.1.2",
     "eslint-plugin-unicorn": "^33.0.1",
+    "eslint-plugin-unused-imports": "^1.1.1",
     "patch-package": "^6.4.7",
-    "typescript": "^4.3.2"
+    "typescript": "^4.3.4"
   },
   "//": "temp rules off bc they seem bugged atm I have no idea why",
   "eslintConfig": {
