+++ global-scoreboard/src/Components/Dashboard/QuickView/QuickView.css
