+++ global-scoreboard/src/Components/Dashboard/Scoreboard.css
