+++ global-scoreboard/src/Components/Dashboard/TableElements/FriendButton.tsx
