+++ global-scoreboard/src/Components/Dashboard/TableElements/PaginationProps.tsx
