+++ global-scoreboard/src/Components/Dashboard/TableElements/PlayerNameCell.css
