+++ global-scoreboard/src/Components/Dashboard/TableElements/PlayerScoreCell.css
