+++ global-scoreboard/src/Components/Dashboard/TableElements/PlayerScoreCell.tsx
@@ -3,9 +3,9 @@ import './PlayerScoreCell.css'
 import { useState } from 'react'
 import { Button } from 'react-bootstrap'
 
-import { renderScoreTable } from 'src/Dashboard/UpdateMessage'
+import { renderScoreTable } from 'src/Components/Dashboard/UpdateMessage'
+import GenericModal from 'src/Components/GenericModal'
 import { apiGet } from 'src/fetchers/Api'
-import GenericModal from 'src/GenericModal'
 import type Player from 'src/Models/Player'
 import type { RunResult } from 'src/Models/UpdateRunnerResult'
