+++ global-scoreboard/src/Components/Dashboard/TableElements/ScoreTitle.tsx
