+++ global-scoreboard/src/Components/GenericModal.tsx
