+++ global-scoreboard/src/Components/NavBar/ScoreboardNavBar.css
@@ -1,0 +1,57 @@
+
+/* Buttons' text in nav-bar should not wrap */
+.navbar .btn {
+  white-space: nowrap;
+}
+
+/* Allow the title text to wrap */
+.navbar-brand {
+  white-space: normal;
+  flex: 1;
+}
+
+.navbar-nav,
+.navbar-text {
+  padding-right: 0.5em;
+}
+
+/* Fixes input group append with Dashlane button */
+.input-group > .form-control+.btn.btn:last-of-type {
+  border-top-right-radius: 0.25rem;
+  border-bottom-right-radius: 0.25rem;
+}
+
+.src-api-key-link .form-control {
+  padding: 0.375rem;
+  text-align: center;
+}
+
+/* Mobile */
+@media screen and (max-width: 397px) {
+  .modal-body .input-group > .btn {
+    width: 100%;
+  }
+  
+  .modal-body .input-group > :first-child {
+    border-bottom-left-radius: 0;
+    margin-left: -1px;
+  }
+  .modal-body .input-group > .form-control {
+    margin-right: 1px;
+  }
+  .modal-body .input-group > .form-control+.btn.btn.btn:last-of-type {
+    border-top-right-radius: 0;
+    border-bottom-left-radius: 0.25rem;
+    margin-top: -1px;
+  }
+
+  .input-group > .input-group-text+.form-control.form-control[name="src-api-key"] {
+    border-top-right-radius: 0.25rem;
+  }
+
+  .src-api-key-link > .input-group > .btn.btn:first-child {
+    border-top-right-radius: 0.25rem;
+    border-bottom-left-radius: 0;
+    margin-bottom: -1px;
+  }
+}
