+++ global-scoreboard/src/Components/NavBar/ScoreboardNavBar.tsx
