--- global-scoreboard/src/NavBar/ScoreboardNavBar.css
@@ -1,75 +1,0 @@
-
-/* Buttons' text in nav-bar should not wrap */
-.navbar .btn {
-  white-space: nowrap;
-}
-
-.navbar-nav,
-.navbar-text {
-  padding-right: 0.5em;
-}
-
-.invalid-feedback {
-  white-space: pre-wrap;
-}
-.input-group-prepend~.invalid-feedback {
-  margin-left: 38px;
-}
-.form-control~.invalid-feedback {
-  padding-left: 0.75rem;
-}
-
-#src-api-key {
-  min-width: 165px;
-}
-
-.input-group > .input-group-append:not(:last-child) > .btn {
-  border-top-right-radius: 0.25rem;
-  border-bottom-right-radius: 0.25rem;
-}
-
-/* Mobile */
-@media screen and (max-width: 382px) {
-  .input-group-prepend > .input-group-text{
-    border-bottom-left-radius: 0;
-    margin-right: -1px;
-  }
-
-  #src-api-key {
-    border-top-right-radius: 0.25rem;
-  }
-
-  .form-group > .input-group > .input-group-append > a.btn {
-    border-top-right-radius: 0;
-    border-bottom-left-radius: 0.25rem;
-    border-bottom-right-radius: 0.25rem;
-    margin-top: -1px;
-  }
-  
-  .src-api-key-link > .input-group > .input-group-prepend,
-  .src-api-key-link > .input-group > .input-group-prepend > a,
-  .src-api-key-link > .input-group > .input-group-append,
-  .src-api-key-link > .input-group > .input-group-append > a,
-  .form-group > .input-group > .input-group-append,
-  .form-group > .input-group > .input-group-append > a {
-    width: 100%
-  }
-
-  .src-api-key-link > .input-group > .input-group-prepend > a {
-    border-top-right-radius: 0.25rem;
-    border-bottom-left-radius: 0;
-    margin-bottom: -1px;
-  }
-
-  .src-api-key-link > .input-group > .input-group-append > a {
-    border-top-right-radius: 0;
-    border-bottom-left-radius: 0.25rem;
-    margin-top: -1px;
-  }
-
-  .input-group-prepend,
-  .input-group-append {
-    margin-left: 0;
-    margin-right: 0;
-  }
-}
