+++ global-scoreboard/src/index.tsx
@@ -1,8 +1,10 @@
 import './index.css'
 
+import { StrictMode } from 'react'
 import ReactDOM from 'react-dom'
+import { BrowserRouter } from 'react-router-dom'
 
-import App from './App'
+import App from './Components/App'
 import * as serviceWorker from './serviceWorker'
 
 // Hack for local network testing
@@ -19,7 +21,15 @@ if ((process.env.REACT_APP_BASE_URL?.includes('127.0.0.1') || process.env.REACT_
     `to ${window.process.env.REACT_APP_BASE_URL}`)
 }
 
-ReactDOM.render(<App />, document.getElementById('root'))
+ReactDOM.render(
+  // Setup static Providers here
+  <StrictMode>
+    <BrowserRouter basename='/global-scoreboard'>
+      <App />
+    </BrowserRouter>
+  </StrictMode>,
+  document.getElementById('root')
+)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
