+++ global-scoreboard/src/utils/Time.ts
@@ -5,6 +5,7 @@ export const secondsToTimeString = (totalSeconds: number) => {
   totalSeconds %= math.SECONDS_IN_HOUR
   const minutes = Math.trunc(totalSeconds / math.SECONDS_IN_MINUTE).toString().padStart(2, '0')
   const seconds = Math.trunc(totalSeconds % math.SECONDS_IN_MINUTE).toString().padStart(2, '0')
+
   return `${hours}:${minutes}:${seconds}`
 }
 
@@ -15,6 +16,7 @@ export const timeStringToSeconds = (timeString: string) => {
   const hour = time[time.length - 3] ?? '0'
   const min = time[time.length - 2] ?? '0'
   const sec = time[time.length - 1]
+
   return +hour * math.SECONDS_IN_HOUR + +min * math.SECONDS_IN_MINUTE + +sec
 }
 
@@ -24,6 +26,7 @@ export const diffDays = (startDate: Date, endDate: Date) => {
   // A day in UTC always lasts 24 hours (unlike in other time formats)
   const start = Date.UTC(startDate.getFullYear(), startDate.getMonth(), startDate.getDate())
   const end = Date.UTC(endDate.getFullYear(), endDate.getMonth(), endDate.getDate())
+
   // so it's safe to divide by 24 hours
   return (end - start) / -math.MS_IN_DAY
 }
