+++ tournament-scheduler/src/Components/LoginForm/SrcApiKeyLink.tsx
@@ -4,7 +4,10 @@ import { useState } from 'react'
 
 const MIN_WIDTH = 9 // SRC name
 
-const SrcApiKeyLink: FC = () => {
+const InputSpan: FC<{ styleProps: unknown }> = ({ styleProps, ...props }) =>
+  <span {...props}>{props.children}</span>
+
+const SrcApiKeyLink = () => {
   const [srcNameInput, setSrcNameInput] = useState('')
   const [linkHover, setStartHover] = useState(false)
   const theme = useTheme()
@@ -24,16 +27,18 @@ const SrcApiKeyLink: FC = () => {
     >www.speedrun.com/</Link>
     <Input
       // divs are not allowed as childs of p
-      components={{ Root: 'span' }}
+      // eslint-disable-next-line react/display-name
+      components={{ Root: InputSpan }}
       name='src-name'
       placeholder='SRC name'
-      // HACK using span as ocmponent looses ::before, so we have to reimplement this
+      // HACK using span as component looses ::before, so we have to reimplement this
       style={{
         borderBottomColor: srcNameInput.length > 0 ? theme.palette.primary.main : theme.palette.text.disabled,
         borderBottomStyle: 'solid',
         borderBottomWidth: '2px',
       }}
       inputProps={{
+        'data-lpignore': 'true',
         style: {
           // eslint-disable-next-line @typescript-eslint/no-magic-numbers
           width: `${srcNameInput.length > 0 ? srcNameInput.length + 0.5 : MIN_WIDTH}ch`,
