+++ tournament-scheduler/src/Models/User.ts
@@ -1,0 +1,5 @@
+type User = {
+  userId: string
+  name: string
+}
+export default User
